Update README.md
Introduction:

This lab will create a class, University Member (UnivMember), which will contain and use the ID and Name classes you built previously, adding a Role.

Instructions:

Create the UnivMember class holding at least:

UnivID

UnivName

Role (Prof, Student or Admin). For now, Role is a string

Each person’s name will be read from a file

The format is the same as before: last,first

Note: The first name is everything to the right of the comma

Please build and use a helper function inside of Name that will remove any trailing blank chars at the end of the first name.